requirements.txt
rich==13.9.4
discord.py==2.4.0
python-dotenv==1.0.1